app/requirements.txt
aniso8601==9.0.1
attrs==21.2.0
blinker==1.4
certifi==2021.5.30
cffi==1.14.3
charset-normalizer==2.0.2
click==7.1.2
cryptography==3.4.7
Flask==1.1.4
Flask-Cors==3.0.10
Flask-Mail==0.9.1
flask-restx==0.4.0
Flask-SQLAlchemy==2.5.1
greenlet==0.4.13
idna==3.2
importlib-metadata==4.6.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==2.0.1
password-strength==0.0.3.post2
phonenumbers==8.12.27
prometheus-client==0.11.0
prometheus-flask-exporter==0.18.2
PyJWT==2.1.0
PyMySQL==0.10.0
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
slugify==0.0.1
SQLAlchemy==1.3.18
SQLAlchemy-Utils==0.37.8
twilio==6.44.1
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==1.0.1
zipp==3.5.0
gunicorn==20.1.0
eventlet==0.30.2